% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_client.R
\docType{data}
\name{ServerClient}
\alias{ServerClient}
\title{ServerClient}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ServerClient
}
\description{
ServerClient
}
\keyword{datasets}

