% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_client.R
\docType{data}
\name{ServerClientDev}
\alias{ServerClientDev}
\title{ServerClientDev}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ServerClientDev
}
\description{
ServerClientDev
}
\keyword{datasets}

