% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\docType{data}
\name{ShinySessionApp}
\alias{ShinySessionApp}
\title{ShinySessionApp}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ShinySessionApp
}
\description{
ShinySessionApp
}
\keyword{datasets}

