% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\docType{data}
\name{ShinySessionDispatcher}
\alias{ShinySessionDispatcher}
\title{ShinySessionDispatcher}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ShinySessionDispatcher
}
\description{
ShinySessionDispatcher
}
\keyword{datasets}

